ga4-table-updated-dataform-run-func/requirements.txt
functions-framework==3.2.0
google-cloud-dataform
pydantic